requirements.txt
flask
google-generativeai
dotenv
streamlit
